Backport 83bce94cc8a7fb45b0604598411fbecc62000dfd
