+++ jdk/test/javax/xml/jaxp/transform/8268457/SurrogateTest2.txt
